server/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
cloudinary==1.44.1
Django==5.2.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.16.1
ecdsa==0.19.1
idna==3.10
Markdown==3.9
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
python-dotenv==1.1.1
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
